refactor: 휴면 해제 페이지 코드 리팩토링 및 디자인 수정
